PCHR-1687: Refactoring and cleaning

+++ uk.co.compucorp.civicrm.hrcore/CRM/HRCore/Upgrader/Steps/1000.php
@@ -1,19 +1,28 @@
 <?php
 
+/**
+ * Class CRM_HRCore_Upgrader_Steps_1000
+ */
 trait CRM_HRCore_Upgrader_Steps_1000 {
 
+  /**
+   * Upgrader to set default Localisation related Settings
+   *
+   * @return bool
+   */
   public function upgrade_1000() {
     $this->up1000_downloadUKEnglish();
     $this->up1000_updateLocalisationSettings();
     $this->up1000_setAvailableCountries();
     $this->up1000_setAvailableProvinces();
-    $this->up1000_setCiviHRTheme();
 
     return TRUE;
   }
 
   /**
    * Downloads en_GB (UK english) localization file
+   *
+   * @return boolean|null
    */
   private function up1000_downloadUKEnglish() {
     $localizationURL = "https://download.civicrm.org/civicrm-l10n-core/mo/en_GB/civicrm.mo";
@@ -21,11 +30,17 @@ trait CRM_HRCore_Upgrader_Steps_1000 {
     global $civicrm_root;
     $downloadPath = "{$civicrm_root}/l10n/en_GB/LC_MESSAGES/";
 
+    if (file_exists("{$downloadPath}civicrm.mo")) {
+      return null;
+    }
+
     if (!is_dir($downloadPath)) {
       mkdir($downloadPath, 0755, true);
     }
 
     file_put_contents($downloadPath, fopen($localizationURL, 'r'));
+
+    return TRUE;
   }
 
   /**
@@ -34,7 +49,7 @@ trait CRM_HRCore_Upgrader_Steps_1000 {
    *      enabled currencies list.
    *   2- setting the default date formats
    *   3- setting the default country to UK
-   *   4- setting the system langagun to UK english (en_GB)
+   *   4- setting the system language to UK english (en_GB)
    */
   private function up1000_updateLocalisationSettings() {
     $settings = [
@@ -46,10 +61,10 @@ trait CRM_HRCore_Upgrader_Steps_1000 {
       'lcMessages' => 'en_GB',
     ];
 
-    // Get UK ID
+    // Get UK Country ID
     $ukCountry = civicrm_api3('Country', 'get', [
-      'return' => array("id"),
-      'iso_code' => "GB",
+      'return' => ['id'],
+      'iso_code' => 'GB',
       'options' => ['limit' => 1],
     ]);
     if (!empty($ukCountry['id'])) {
@@ -65,7 +80,7 @@ trait CRM_HRCore_Upgrader_Steps_1000 {
 
     foreach ($currenciesToEnable as $currency) {
       civicrm_api3('OptionValue', 'create', [
-        'option_group_id' => "currencies_enabled",
+        'option_group_id' => 'currencies_enabled',
         'label' => $currency[0],
         'value' => $currency[1],
         'is_default' => $currency[2],
@@ -75,48 +90,43 @@ trait CRM_HRCore_Upgrader_Steps_1000 {
   }
 
   /**
-   * Sets Available Countries
+   * Sets Available Countries to 'all countries'
    */
   private function up1000_setAvailableCountries() {
-    $countriesList = civicrm_api3('Country', 'get', array(
+    $countriesList = civicrm_api3('Country', 'get',[
       'sequential' => 1,
-      'return' => array("id"),
-      'options' => array('limit' => 0),
-    ));
+      'return' => ['id'],
+      'options' => ['limit' => 0],
+    ]);
 
-    $countriesIDs = array_column($countriesList['values'], 'id');
-    unset($countriesList);
+    if (!empty($countriesList['values'])) {
+      $countriesIDs = array_column($countriesList['values'], 'id');
+      unset($countriesList);
 
-    civicrm_api3('Setting', 'create', array(
-      'countryLimit' => $countriesIDs,
-    ));
+      civicrm_api3('Setting', 'create', [
+        'countryLimit' => $countriesIDs,
+      ]);
+    }
   }
 
   /**
-   * Sets Available Provinces
+   * Sets Available Provinces to 'all provinces'
    */
   private function up1000_setAvailableProvinces() {
+    // ToDo : Fetch Provinces via API after upgrading to newer civicrm version since it is not available on civicrm 4.7.9
     $tableName = CRM_Core_DAO_StateProvince::getTableName();
-
-    $provincesIDs = [];
     $query = CRM_Core_DAO::executeQuery("SELECT id FROM {$tableName}");
 
+    $provincesIDs = [];
     while($query->fetch()) {
       $provincesIDs[] = $query->id;
     }
 
-    civicrm_api3('Setting', 'create', array(
-      'provinceLimit' => $provincesIDs,
-    ));
-  }
-
-  /**
-   * Sets CiviHR theme by updating the custom CSS URL
-   */
-  private function up1000_setCiviHRTheme() {
-    civicrm_api3('Setting', 'create', [
-      'customCSSURL' => '[civicrm.root]/tools/extensions/civihr/org.civicrm.bootstrapcivicrm/css/custom-civicrm.css',
-    ]);
+    if (!empty($provincesIDs)) {
+      civicrm_api3('Setting', 'create', [
+        'provinceLimit' => $provincesIDs,
+      ]);
+    }
   }
 
 }
